fix typo in name change warning
fix comparison in componentDidUpdate to correctly expand named
submission

fix text of naming conflict

clear submissionToUpdate when submissionType changes to "new"

+++ src/Components/DataSubmission/NameChangeWarning.js
@@ -48,7 +48,7 @@ const LongNameWarning = (props) => {
   return (
     <Typography variant="body1" className={cl.title}>
       <ErrorOutline className={cl.warningIcon} />{' '}
-      Short Name will change from {' '}
+      Long Name will change from {' '}
       <span className={cl.bright}>{data.originalLongName}</span> to {' '}
       <span className={cl.bright}>{data.longName}</span>.
     </Typography>
